+++ src/components/base/CollectionModal/ModalHeader.tsx
@@ -1,0 +1,3 @@
+import { styled } from "@mui/material";
+
+export const ModalHeader = styled('div')``
